feat: implement seccomp SECCOMP_RET_LOG and RET_TRAP with audit logging (Item 69)
Implementation of SECCOMP_RET_LOG (log + allow), SECCOMP_RET_TRAP (send SIGSYS)
and enhanced audit logging for all seccomp filter actions.

Changes:
- seccomp.h: Add SECCOMP_RET_LOG, SECCOMP_RET_TRAP constants
- seccomp.h: Add seccomp_evaluate_syscall() with full argument logging
- seccomp.h: Add seccomp_send_sigsys() for RET_TRAP action
- seccomp.c: Implement seccomp_evaluate_syscall() with per-action audit logging
- seccomp.c: Implement RET_LOG (log then allow), RET_TRAP (send SIGSYS signal)
- seccomp.c: Implement RET_KILL audit reporting with process info
- seccomp.c: Add kallsyms-based symbol lookup in audit messages
- seccomp.c: Validate filter actions in seccomp_add_rule()
- syscall.c: Update dispatch to pass syscall args to seccomp evaluator
- syscall.c: Handle RET_TRAP by sending SIGSYS, RET_KILL by killing process
- .gitignore: Add build_check/ to patterns

Previously seccomp only returned allow/deny with no logging. Now audit
records include syscall number + arguments, process PID/name, RIP with
symbol resolution, and the action taken.

+++ src/drivers/usb_msc.c
@@ -435,7 +435,7 @@ int usb_msc_init(void)
     /* SET_ADDRESS (standard SETUP to ep0, device address = 1) */
     int rc = usb_control(0x00, 0x05, g_dev_addr, 0, 0, (void *)0);
     if (rc < 0) {
-        kprintf("  USB MSC: SET_ADDRESS failed (%d)\n", (unsigned long)rc);
+        kprintf("  USB MSC: SET_ADDRESS failed (%ld)\n", (unsigned long)rc);
         return rc;
     }
     busy_wait(50000);
@@ -445,7 +445,7 @@ int usb_msc_init(void)
     if (!desc) return -4;
     rc = usb_control(0x80, 0x06, 0x0100, 0, 18, desc);
     if (rc < 0) {
-        kprintf("  USB MSC: GET_DESCRIPTOR failed (%d)\n", (unsigned long)rc);
+        kprintf("  USB MSC: GET_DESCRIPTOR failed (%ld)\n", (unsigned long)rc);
         pmm_free_frame(VIRT_TO_PHYS((unsigned long)desc));
         return rc;
     }
@@ -455,7 +455,7 @@ int usb_msc_init(void)
     /* SET_CONFIGURATION 1 */
     rc = usb_control(0x00, 0x09, 1, 0, 0, (void *)0);
     if (rc < 0) {
-        kprintf("  USB MSC: SET_CONFIGURATION failed (%d)\n", (unsigned long)rc);
+        kprintf("  USB MSC: SET_CONFIGURATION failed (%ld)\n", (unsigned long)rc);
         return rc;
     }
 
@@ -463,19 +463,19 @@ int usb_msc_init(void)
     rc = usb_control(0x21, 0xFF, 0, 0, 0, (void *)0);
     if (rc < 0) {
         /* Non-fatal — some devices ignore it */
-        kprintf("  USB MSC: BOT reset warning (%d)\n", (unsigned long)rc);
+        kprintf("  USB MSC: BOT reset warning (%ld)\n", (unsigned long)rc);
     }
 
     /* READ CAPACITY to discover drive size */
     rc = scsi_read_capacity(&g_max_lba);
     if (rc < 0) {
-        kprintf("  USB MSC: READ CAPACITY failed (%d)\n", (unsigned long)rc);
+        kprintf("  USB MSC: READ CAPACITY failed (%ld)\n", (unsigned long)rc);
         return rc;
     }
 
-    kprintf("  USB MSC: %u sectors (%u MB)\n",
+    kprintf("  USB MSC: %lu sectors (%llu MB)\n",
             (unsigned long)(g_max_lba + 1),
-            (uint64_t)((g_max_lba + 1) / 2048));
+            (unsigned long long)((g_max_lba + 1) / 2048));
 
     /* Register with block device layer */
     blockdev_register_legacy(BLOCKDEV_USB0, "usb0",
